Add files via upload
The Publisher wanted the number of page views to be visible in the footer of every page. This uses PHP and Google's analytics API to pull all the pageviews for the site and find the one relevent to the current page. A shortcode is then created to be put in the site footer to display the total number of pageviews a page has.